Draft service layer
